Update Min-Max_Guide.md
Change many, but not all, directions to compass directions. Some judgment was used accounting for context. Generally map-scale vs person-scale or outdoors vs indoors or map vs screen. I looked for a good place to put a style guide for this, but I really wasn't comfortable with any place. The closest would be in Route Overview.