Update Min-Max_Guide.md
1. There is potential wasted time waiting for Emily to come from her room that can be used by meeting Jas a little later and further west if needed. Thus planning a clockwise trip (if any) around the lake may be better. This doesn't add travel time since we would have had to double north then south at the end anyway to catch Jas and Haley. Coming from the Secret Woods area at the end makes meeting Jas and Haley more natural and puts Emily in her living room at the risk of chasing Penny toward Marnie.

This worked a lot better than counter-clockwise (which just wasn't working for me) when I tested it. In fact, it seems that the optimal route may be to barely catch Haley before she gets too far south. Going out of the way for Marnie is just more weeds that aren't killed in the faForest, and there are plenty to kill near where Haley stops.

2. Against the event that Spring 1 goes financially poorly (no artifact and scant forage), should the guide recommend as edited?