Spring 1 Update Min-Max_Guide.md (#40)
* Spring 1 Update Min-Max_Guide.md

Small changes. 

It is not mentioned what farm to choose so I assume it's the standard one. There seem to be upsides in choosing four corners but I'm keen to hear thoughts on why that might be a bad idea.

* Spring 1 Update Min-Max_Guide.md, punctuation, grammar

I went ahead and addressed the punctuation and grammar. These resulted in more edits where I originally wanted to make minimal changes.

* addressing comments

* minor

* small changes

* last?

* quest journal on day one

* day 2 changes

* minor

* Update Min-Max_Guide.md

* Update Min-Max_Guide.md

* Update Min-Max_Guide.md

* Update Min-Max_Guide.md

---------

Co-authored-by: James <5511220+Zamiell@users.noreply.github.com>